+++ docs/misc/changelog.rst
@@ -3,17 +3,20 @@
 Changelog
 ==========
 
-Pre-Release 0.4.0a0 (WIP)
+Pre-Release 0.4.0 (2020-02-14)
 ------------------------------
 
+**Proper pre-processing, independent save/load for policies**
 
 Breaking Changes:
 ^^^^^^^^^^^^^^^^^
 - Removed CEMRL
+- Model saved with previous versions cannot be loaded (because of the pre-preprocessing)
 
 New Features:
 ^^^^^^^^^^^^^
-- Add support for Discrete observation spaces
+- Add support for ``Discrete`` observation spaces
+- Add saving/loading for policy weights, so the policy can be used without the model
 
 Bug Fixes:
 ^^^^^^^^^^
@@ -26,6 +29,8 @@ Others:
 ^^^^^^^
 - Refactor handling of observation and action spaces
 - Refactored features extraction to have proper preprocessing
+- Refactored action distributions
+
 
 Documentation:
 ^^^^^^^^^^^^^^